fix(redb_cold_storage): remove unnecessary cast to fix clippy warning
Removed unnecessary 'i as i64' cast in test_multiple_overwrites_same_version.
The type is already inferred as i64 from context.

Fixes clippy::unnecessary_cast warning.

+++ src/storage/redb_cold_storage.rs
@@ -1798,7 +1798,7 @@ mod tests {
         // Write same version multiple times
         for i in 0..5 {
             let props = PropertyMapBuilder::new()
-                .insert("iteration", i as i64)
+                .insert("iteration", i)
                 .build();
 
             let version = NodeVersion::new_anchor(
